// sipfactel_front/src/app/models/resoluciones/ResolucionFacturas.ts
export class ResolucionFacturas {
 llave: number;
 sucursal: number;
 compania: number;
 numeroResolucion: number;
 prefijo: string;
 fechaInicial: string;
 fechaFinal: string;
 numeroInicial: number;
 numeroFinal: number;
 estado: string;
 trTipoId: number;
 vigencia: boolean;
 contador: number;
 producto: string;
}